moved assets out of dtl and flattened the module hierarchy
