+++ hatch/src/cli/commands/build.rs
@@ -4,8 +4,8 @@ use cli::commands::Command;
 use cli::commands::ops::ProjectOps;
 use yaml;
 use project::Project;
-use dtl::asset::{ ProjectAssetBuilder, TupKind, print_file_contents };
-use dtl::tup::Asset;
+use assets::{ Asset, TupKind, print_file_contents };
+use assets::builder::Builder as AssetBuilder;
 
 use cli::commands::PROJECT_NAMES;
 
@@ -69,7 +69,7 @@ impl<'command> Command<'command> for Build {
 
     match result[0] {
       Ok(ref project) => {
-        let mut asset_builder = ProjectAssetBuilder::from(project);
+        let mut asset_builder = AssetBuilder::from(project);
         for asset in asset_builder.assets() {
           println!("{}\n", asset.contents());
         }
