+++ hatch/src/dtl/mod.rs
@@ -1,3 +1,1 @@
-pub mod asset;
 pub mod cli;
-pub mod tup;
